removed operators for automatically mapping optionals in favor of helper functions
