+++ Pipeline/Consumable/ConsumableOperators.swift
@@ -46,29 +46,6 @@ public func |> <U, NewOutput>(lhs: ConsumablePipeline<U>, rhs: U throws -> NewOu
     return lhs.then(resultFunction)
 }
 
-// optional chaining
-
-public func |> <S: ConsumableType, V, NewOutput where S.OutputType == Optional<V>>(lhs: S, rhs: V -> NewOutput) -> ConsumablePipeline<NewOutput?>  {
-    
-    let transformer = optionalMap(rhs)
-    
-    return ConsumablePipeline(head: lhs).then(transformer)
-}
-
-public func |> <S: ConsumableType, V, T: TransformerType where S.OutputType == Optional<V>, V == T.InputType>(lhs: S, rhs: T) -> ConsumablePipeline<T.OutputType?>  {
-    
-    let transformer = optionalMap(rhs)
-    
-    return ConsumablePipeline(head: lhs).then(transformer)
-}
-
-public func |> <U, NewOutput>(lhs: ConsumablePipeline<U?>, rhs: U -> NewOutput) -> ConsumablePipeline<NewOutput?>  {
-    
-    let mappedTransform = optionalMap(rhs)
-    
-    return lhs.then(mappedTransform)
-}
-
 // finally
 
 public func |> <S: ConsumableType, C: ConsumerType where S.OutputType == C.InputType>(lhs: S, rhs: C) -> Pipeline  {
