+++ PipelineTests/HelperTests.swift
@@ -165,110 +165,6 @@ class HelperTests: XCTestCase {
         pipe.produce()
     }
     
-    func testSwallowError() {
-        
-        let result: Result<String> = .Error(MockError())
-        
-        let pipe = ValueProducer(result)
-            |> swallowError(log: "found error")
-            |> { _ in XCTAssert(false) }
-        
-        pipe.produce()
-    }
-    
-    func testSwallowErrorSuccess() {
-        
-        let result: Result<String> = .Success("abc")
-        
-        let expt = expectationWithDescription("error")
-        
-        let pipe = ValueProducer(result)
-            |> swallowError(log: "found error")
-            |> { (str: String) in
-                
-                XCTAssert(str == "abc")
-                expt.fulfill()
-        }
-        
-        pipe.produce()
-        
-        waitForExpectationsWithTimeout(0.1, handler: nil)
-    }
-    
-    func testLogError() {
-        
-        let result: Result<String> = .Error(MockError())
-        
-        let pipe = ValueProducer(result)
-            |> logError("found error")
-            |> { _ in XCTAssert(false) }
-        
-        pipe.produce()
-    }
-    
-    func testLogErrorSuccess() {
-        
-        let result: Result<String> = .Success("abc")
-        
-        let expt = expectationWithDescription("error")
-        
-        let pipe = ValueProducer(result)
-            |> logError("found error")
-            |> { (str: String) in
-                
-                XCTAssert(str == "abc")
-                expt.fulfill()
-        }
-        
-        pipe.produce()
-        
-        waitForExpectationsWithTimeout(0.1, handler: nil)
-    }
-    
-    func testOnError() {
-        
-        let result: Result<String> = .Error(MockError())
-        
-        let expt = expectationWithDescription("error")
-        
-        let pipe = ValueProducer(result) |> onError() { err in
-                
-                XCTAssert(err is MockError)
-                
-                expt.fulfill()
-            
-        } |> { (str: String) in
-            
-            XCTFail()
-        }
-        
-        pipe.produce()
-        
-        waitForExpectationsWithTimeout(0.1, handler: nil)
-    }
-    
-    func testOnErrorSuccess() {
-        
-        let result: Result<String> = .Success("success")
-        
-        let expt = expectationWithDescription("error")
-        
-        let pipe = ValueProducer(result) |> onError() { err in
-            
-            XCTFail()
-            
-        } |> { (str: String) in
-                
-            XCTAssert(str == "success")
-            
-            expt.fulfill()
-        }
-        
-        pipe.produce()
-        
-        waitForExpectationsWithTimeout(0.1, handler: nil)
-    }
-    
     func testOnNil() {
         
         let optional: String? = nil
@@ -282,6 +178,26 @@ class HelperTests: XCTestCase {
         } |> { (str: String) in
                 
             XCTFail()
+        }
+        
+        pipe.produce()
+        
+        waitForExpectationsWithTimeout(0.1, handler: nil)
+    }
+    
+    func testResolveNil() {
+        
+        let optional: String? = nil
+        
+        let expt = expectationWithDescription("nil")
+        
+        let pipe = ValueProducer(optional)
+            |> resolveNil() { return "abc" }
+            |> { (str: String) in
+                
+            XCTAssert(str == "abc")
+            
+            expt.fulfill()
         }
         
         pipe.produce()
